prod/requirements.txt
torch
torchvision
streamlit
facenet-pytorch
Pillow
numpy
scikit-learn
matplotlib